🌈 Update: add prometheus url
