+++ src/pages/creator/CreatorPage.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, useContext, useEffect, useState } from 'react';
 import { ContextProps, SidebarMenu } from '../../types/Props.tsx';
-import Sidebar from '../../components/Sidebar.tsx';
+import Sidebar from '../../layout/Sidebar.tsx';
 import IconList from './IconList.tsx';
 import IconDetail from './IconDetail.tsx';
 import IconRegister from './IconRegister.tsx';
